Add new CreateHttpApiRequest with TActionResponse Fix CodeQL alerts Test
